Migrated to VS2013.
Beware, as, this version is fully Unicode, it would probably need some
changes in the Autoit wrapper before it can be used.
LoadFromFile is now implemented, but it seems to introduce some memory
corruption: searchs work fine as far as I can tell after loading, but it
failed to save the loaded bitmap into another file.